requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
mongoengine
passlib[bcrypt] 
python-jose[cryptography]
email-validator
python-multipart
Pillow